ai-agent/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
langchain==0.0.220
openai==0.27.8
python-dotenv==1.0.0
pydantic==1.10.9
requests==2.31.0
websockets==11.0.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
